+++ internal/browserplugin/register.go
@@ -1,7 +1,0 @@
-package browserplugin
-
-import "github.com/devlikebear/tars/internal/plugin"
-
-func init() {
-	plugin.RegisterBuiltin(&Plugin{})
-}